--- CREDITS.md
@@ -1,25 +1,0 @@
-### Contributors
-
-The Zhongwen Chrome extension was written by Christian Schiller with
-contributions from the following people:
-
-* Todd Rudick (http://rikaixul.mozdev.org/)
-
-* Jonathan Zarate (http://www.polarcloud.com/)
-
-* Erek Speed (http://code.google.com/p/rikaikun/)
-
-* Lenoard Lausen (https://github.com/leezu)
-
-Furthermore, Zhongwen would not be possible without the work of all the editors
-who put together, and constantly improve and update, the
-CC-CEDICT dictionary: https://www.mdbg.net/chinese/dictionary?page=cedict
-
-I also want to thank John Pasden from AllSet Learning
-(https://www.allsetlearning.com/about/) for creating, and suggesting
-Zhongwen be integrated with, the Chinese Grammar Wiki.
-(https://resources.allsetlearning.com/chinese/grammar/)
-
-Thanks also to Nick Winter from Skritter for his assistance with
-setting up the API for integrating Zhongwen with Skritter.
-(https://skritter.com/)
