--- background.html
@@ -1,7 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <script type="module" src="./dict.js"></script>
-        <script type="module" src="./background.js"></script>
-    </head>
-</html>
