--- css/wordlist.css
@@ -1,3 +1,0 @@
-body {
-    margin: 20px;
-}
