--- data/grammarKeywordsMin.json
@@ -1,1 +1,0 @@
-{"啊":1,"吧":1,"倍":1,"呗":1,"被":1,"本":1,"本来":1,"别":1,"别人":1,"别说":1,"边":1,"不了":1,"半":1,"办法":1,"白":1,"白白":1,"必":1,"毕竟":1,"并":1,"并且":1,"不":1,"不是":1,"不像":1,"不怎么":1,"不知道":1,"不比":1,"不曾":1,"不错":1,"不得不":1,"不得了":1,"不断":1,"不但":1,"不到":1,"不动":1,"不懂":1,"不过":1,"不管":1,"不仅":1,"不可":1,"不愧":1,"不愧是":1,"不论":1,"不起":1,"不然":1,"不如":1,"不停":1,"不屑":1,"不下":1,"不行":1,"不要":1,"不用":1,"不着":1,"不只":1,"八":1,"包括":1,"把":1,"百":1,"比":1,"比起":1,"比起来":1,"彼此":1,"比较":1,"曾经":1,"趁":1,"成":1,"差":1,"差很多":1,"差不多":1,"差点":1,"差点没":1,"常常":1,"朝":1,"除非":1,"除了":1,"出":1,"出来":1,"出去":1,"错":1,"才":1,"从":1,"从而":1,"从不":1,"从来":1,"从没":1,"的":1,"的时候":1,"得":1,"得多":1,"的话":1,"得看":1,"等":1,"等等":1,"等到":1,"等于":1,"点":1,"点钟":1,"对":1,"对方":1,"对于":1,"多":1,"多长时间":1,"多亏":1,"多了":1,"多少":1,"但":1,"但是":1,"倒":1,"到":1,"到底":1,"地":1,"第":1,"动":1,"当":1,"当中":1,"都":1,"懂":1,"二":1,"而":1,"而且":1,"而是":1,"而言":1,"而已":1,"非":1,"非得":1,"非要":1,"分之":1,"凡是":1,"方面":1,"反而":1,"反正":1,"否":1,"否则":1,"个":1,"格":1,"给":1,"跟":1,"根本":1,"更":1,"根据":1,"格外":1,"怪":1,"怪不得":1,"过":1,"过来":1,"过去":1,"关于":1,"果然":1,"够":1,"够了":1,"该":1,"干脆":1,"刚":1,"刚才":1,"刚刚":1,"和":1,"何必":1,"何况":1,"很":1,"很是":1,"合适":1,"化":1,"坏":1,"还原":1,"会":1,"回":1,"回来":1,"回去":1,"或":1,"或者":1,"恢复":1,"害":1,"害得":1,"号":1,"还":1,"还好":1,"还是":1,"还有":1,"候":1,"后":1,"后来":1,"互相":1,"乎":1,"忽然":1,"好":1,"好不容易":1,"好容易":1,"好像":1,"借":1,"结果":1,"见":1,"鉴于":1,"叫":1,"假如":1,"就":1,"就是":1,"就算":1,"兼":1,"将":1,"将会":1,"将要":1,"究竟":1,"简直":1,"角":1,"假使":1,"九":1,"既":1,"近":1,"进":1,"竟":1,"竟然":1,"尽量":1,"进来":1,"进去":1,"进一步":1,"及":1,"即使":1,"据":1,"几乎":1,"经常":1,"经过":1,"居然":1,"几":1,"尽管":1,"仅仅":1,"既然":1,"可":1,"肯定":1,"可是":1,"可以":1,"块":1,"快":1,"看":1,"看起来":1,"看来":1,"开":1,"恐怕":1,"了":1,"类":1,"谅":1,"连":1,"六":1,"两":1,"来":1,"来说":1,"来着":1,"来自":1,"另外":1,"离":1,"老":1,"老是":1,"里":1,"礼拜":1,"吗":1,"嘛":1,"每":1,"没":1,"没有":1,"毛":1,"名":1,"名义":1,"莫":1,"莫非":1,"莫过于":1,"呢":1,"内":1,"那个":1,"能":1,"年":1,"None":1,"那":1,"那么":1,"难":1,"难道":1,"难怪":1,"难免":1,"宁可":1,"宁肯":1,"宁愿":1,"哪里":1,"哪怕":1,"哪儿":1,"偏":1,"凭":1,"平常":1,"平时":1,"且":1,"前":1,"千":1,"千万":1,"却":1,"确定":1,"全":1,"去":1,"七":1,"其中":1,"岂":1,"起":1,"岂不是":1,"起来":1,"请":1,"任何":1,"人家":1,"让":1,"然":1,"然后":1,"日":1,"容易":1,"如":1,"如果":1,"谁":1,"什么":1,"什么时候":1,"甚至":1,"上":1,"上来":1,"上去":1,"是":1,"势必":1,"似的":1,"是否":1,"适合":1,"十":1,"时":1,"时候":1,"受":1,"顺便":1,"少":1,"少不了":1,"使":1,"算":1,"岁":1,"随着":1,"虽然":1,"所":1,"所谓":1,"所以":1,"所有":1,"四":1,"似乎":1,"三":1,"死":1,"太":1,"通过":1,"突然":1,"倘若":1,"挺":1,"往":1,"谓":1,"为":1,"未曾":1,"为了":1,"为什么":1,"外":1,"万":1,"万一":1,"完":1,"无":1,"无非":1,"无关":1,"无论":1,"五":1,"些":1,"下":1,"下来":1,"像":1,"向":1,"向来":1,"下去":1,"先":1,"相比":1,"相反":1,"显得":1,"想":1,"姓":1,"性":1,"幸好":1,"幸亏":1,"星期":1,"也":1,"月":1,"越":1,"越来越":1,"元":1,"原来":1,"远":1,"要":1,"要不是":1,"要不":1,"要看":1,"要么":1,"要命":1,"要是":1,"亿":1,"一点":1,"又":1,"由":1,"由此":1,"由于":1,"于":1,"于是":1,"一":1,"一方面":1,"一边":1,"一点儿":1,"一旦":1,"一定":1,"一度":1,"一律":1,"一面":1,"因此":1,"应该":1,"因为":1,"一起":1,"一时":1,"一些":1,"一下":1,"一向":1,"一下子":1,"一样":1,"一直":1,"一再":1,"以":1,"以便":1,"以后":1,"以及":1,"已经":1,"以来":1,"以免":1,"以前":1,"以上":1,"以为":1,"以外":1,"以下":1,"有":1,"有的":1,"有的是":1,"有点":1,"有点儿":1,"有关":1,"与":1,"与否":1,"与其":1,"则":1,"怎么":1,"怎么样":1,"着":1,"这":1,"这个":1,"这么":1,"真":1,"正":1,"正好":1,"正在":1,"这样":1,"至于":1,"值得":1,"直到":1,"住":1,"之":1,"之后":1,"之间":1,"之类的":1,"之前":1,"之所以":1,"之一":1,"中":1,"中间":1,"终于":1,"周":1,"只":1,"只好":1,"只要":1,"只有":1,"最":1,"最好":1,"做主":1,"再":1,"在":1,"极了":1,"再也不":1,"再不":1,"在内":1,"再说":1,"再三":1,"在于":1,"自从":1,"足":1,"足足":1,"早":1,"早在":1,"总":1,"总得":1,"总归":1,"总是":1,"总算":1,"总之":1,"按照":1}
