--- data/vocabularyKeywordsMin.json
@@ -1,1 +1,0 @@
-{"增加":1,"增长":1,"变":1,"遍":1,"变化":1,"表达":1,"表示":1,"办法":1,"办公室":1,"帮":1,"帮忙":1,"帮助":1,"本地":1,"摆":1,"保存":1,"保护":1,"保留":1,"创作":1,"创造":1,"穿":1,"常常":1,"藏":1,"曾经":1,"次":1,"躲":1,"大概":1,"带":1,"戴":1,"大约":1,"得到":1,"度假":1,"当地":1,"担心":1,"等":1,"等等":1,"都":1,"打算":1,"打听":1,"放":1,"放假":1,"放松":1,"肥":1,"附近":1,"发":1,"发达":1,"方法":1,"发展":1,"关心":1,"拐":1,"管":1,"干":1,"高兴":1,"搁":1,"根据":1,"工夫":1,"公司":1,"工作":1,"改":1,"改变":1,"改进":1,"改善":1,"感觉":1,"感激":1,"感谢":1,"搞":1,"话题":1,"回答":1,"回复":1,"花":1,"害怕":1,"很多":1,"见":1,"见面":1,"建议":1,"教师":1,"旧":1,"家":1,"将来":1,"家庭":1,"解释":1,"觉得":1,"寄":1,"计划":1,"进步":1,"据说":1,"经常":1,"经历":1,"经验":1,"快乐":1,"看":1,"控制":1,"开心":1,"了解":1,"临时":1,"老":1,"老师":1,"理解":1,"目标":1,"目的":1,"美丽":1,"满意":1,"满足":1,"年纪":1,"年龄":1,"年轻":1,"年青":1,"弄":1,"漂亮":1,"怕":1,"胖":1,"普遍":1,"普通":1,"全":1,"气候":1,"气温":1,"期待":1,"轻松":1,"期望":1,"认识":1,"认为":1,"说明":1,"上班":1,"时间":1,"商量":1,"收到":1,"少":1,"天气":1,"推荐":1,"提高":1,"题目":1,"听说":1,"讨论":1,"未来":1,"问":1,"维护":1,"休假":1,"想法":1,"小":1,"希望":1,"许多":1,"严肃":1,"严重":1,"意见":1,"意思":1,"意义":1,"用":1,"愉快":1,"已经":1,"以为":1,"转":1,"着急":1,"知道":1,"周围":1,"主题":1,"主意":1,"作":1,"做":1,"在乎":1,"暂时":1,"按照":1}
