--- help.html
@@ -1,110 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <title>Zhongwen Help</title>
-    <meta charset="utf-8" />
-    <link rel="shortcut icon" href="images/zhongwen.png" />
-    <link rel="stylesheet" type="text/css" href="css/bootstrap.min.css" />
-    <link>
-</head>
-<body>
-    <div class="container">
-        <h1 class="my-3">Zhongwen Help</h1>
-
-        <h3>Keyboard shortcuts</h3>
-        <table class="table table-sm">
-            <tr>
-                <th class="pt-3" colspan="2">Navigation</th>
-            </tr>
-            <tr>
-                <td><kbd>N</kbd></td>
-                <td>Next word</td>
-            </tr>
-            <tr>
-                <td><kbd>B</kbd></td>
-                <td>Previous character ("back")</td>
-            </tr>
-            <tr>
-                <td><kbd>M</kbd></td>
-                <td>Next character</td>
-            </tr>
-            <tr>
-                <th class="pt-3" colspan="2">Moving the pop-up window</th>
-            </tr>
-            <tr>
-                <td><kbd>A</kbd></td>
-                <td>Move the pop-up location between three different locations</td>
-            </tr>
-            <tr>
-                <td><kbd>Y</kbd></td>
-                <td>Move pop-up window down</td>
-            </tr>
-            <tr>
-                <td><kbd>X</kbd></td>
-                <td>Move pop-up window up</td>
-            </tr>
-            <tr>
-                <th class="pt-3" colspan="2">Working with text</th>
-            </tr>
-            <tr>
-                <td><kbd>C</kbd></td>
-                <td>Copy translation to clipboard</td>
-            </tr>
-            <tr>
-                <td><kbd>Ctrl</kbd> + <kbd>C</kbd></td>
-                <td>Just copy the selected Chinese characters (without translation) to the clipboard</td>
-            </tr>
-            <tr>
-                <th class="pt-3" colspan="2">Learning words and characters</th>
-            </tr>
-            <tr>
-                <td><kbd>R</kbd></td>
-                <td>Remember word by adding it to the internal word list</td>
-            </tr>
-            <tr>
-                <td><kbd>S</kbd></td>
-                <td>Add word to Skritter queue</td>
-            </tr>
-            <tr>
-                <td><kbd>Alt</kbd> + <kbd>W</kbd></td>
-                <td>Show the internal word list</td>
-            </tr>
-            <tr>
-                <th class="pt-3" colspan="2">External dictionaries</th>
-            </tr>
-            <tr>
-                <td><kbd>Alt</kbd> + <kbd>1</kbd></td>
-                <td><a href="https://english.dict.naver.com/english-chinese-dictionary/#/main" target="_blank">LINE Dict</a></td>
-            </tr>
-            <tr>
-                <td><kbd>Alt</kbd> + <kbd>2</kbd></td>
-                <td><a href="https://forvo.com/" target="_blank">Forvo</a></td>
-            </tr>
-            <tr>
-                <td><kbd>Alt</kbd> + <kbd>3</kbd></td>
-                <td><a href="https://dict.cn" target="_blank">Dict.cn</a></td>
-            </tr>
-            <tr>
-                <td><kbd>Alt</kbd> + <kbd>4</kbd></td>
-                <td><a href="https://www.iciba.com" target="_blank">iCIBA</a></td>
-            </tr>
-            <tr>
-                <td><kbd>Alt</kbd> + <kbd>5</kbd></td>
-                <td><a href="https://mdbg.net" target="_blank">MDBG</a></td>
-            </tr>
-            <tr>
-                <td><kbd>Alt</kbd> + <kbd>6</kbd></td>
-                <td><a href="https://reverso.com" target="_blank">Reverso</a></td>
-            </tr>
-            <tr>
-                <td><kbd>Alt</kbd> + <kbd>7</kbd></td>
-                <td><a href="https://moedict.tw" target="_blank">MoE Dict</a></td>
-            </tr>
-            <tr>
-                <td><kbd>T</kbd></td>
-                <td><a href="https://tatoeba.org" target="_blank">Tatoeba</a></td>
-            </tr>
-        </table>
-    </div>
-</body>
-</html>
