--- manifest.json
@@ -1,55 +1,0 @@
-{
-  "manifest_version": 2,
-  "name": "Zhongwen: Chinese-English Dictionary",
-  "short_name": "Zhongwen",
-  "version": "5.16.0",
-  "author": "Christian Schiller",
-  "homepage_url": "https://github.com/cschiller/zhongwen",
-  "description": "Great tool for learning Chinese. Includes links to Chinese Grammar Wiki. Supports adding words to Skritter.",
-  "icons": {
-    "16": "images/zhongwen16.png",
-    "48": "images/zhongwen48.png",
-    "128": "images/zhongwen.png"
-  },
-  "browser_action": {
-    "default_icon": "images/zhongwen48.png",
-    "default_title": "Zhongwen Chinese Popup Dictionary"
-  },
-  "background": {
-    "page": "background.html"
-  },
-  "content_scripts": [
-    {
-      "matches": [
-        "<all_urls>"
-      ],
-      "js": [
-        "js/jquery-3.3.1.min.js",
-        "js/zhuyin.js",
-        "content.js"
-      ],
-      "css": [
-        "css/content.css"
-      ],
-      "all_frames": true
-    }
-  ],
-  "options_ui": {
-    "page": "options.html",
-    "open_in_tab": true
-  },
-  "permissions": [
-    "contextMenus",
-    "tabs"
-  ],
-  "commands": {
-    "_execute_browser_action": {
-      "description": "Enable/Disable Zhongwen"
-    }
-  },
-  "web_accessible_resources": [
-    "css/*",
-    "js/*",
-    "images/*"
-  ]
-}
