+++ src/background.html
@@ -1,0 +1,7 @@
+<!DOCTYPE html>
+<html>
+    <head>
+        <script type="module" src="./dict.js"></script>
+        <script type="module" src="./background.js"></script>
+    </head>
+</html>
