+++ src/css/wordlist.css
@@ -1,0 +1,3 @@
+body {
+    margin: 20px;
+}
